backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
boto3==1.34.69
pdfplumber==0.10.3
groq==0.9.0
ddtrace==2.7.0
datadog==0.49.1
Pillow==10.2.0
gunicorn==21.2.0
uvicorn==0.27.1
httpx==0.27.0
geopy==2.4.1
Faker==24.2.0
numpy==1.26.4
pandas==2.2.1
